+++ R/conditional_effects_bgm.R
@@ -68,10 +68,10 @@
 #' # Population average distance curve
 #' conditional_effects_bgm(model, deriv = 0, re_formula = NA)
 #' 
+#' \donttest{
 #' # Individual-specific distance curves
 #' conditional_effects_bgm(model, deriv = 0, re_formula = NULL)
 #' 
-#' \donttest{
 #' # Population average velocity curve
 #' conditional_effects_bgm(model, deriv = 1, re_formula = NA)
 #'