+++ NEWS.md
@@ -1,4 +1,4 @@
 # bsitar 0.1.0
 
 * First release on CRAN.
-* Added a `NEWS.md` file to track future changes to the package.
+* Added a `NEWS.md` file to track future changes.